update: refactor code from main
